All tests passing. Introduced J methods for escapement.
